script_comp: fix undesired resman container set mutations
This commit modifies the method in which nwn_script_comp searches ResMan
to deliver the desired ResRef.  Previous behavior allowed for the
ResMan.containers set to be mutatated in
`nwn_script_comp::serviceRmDemand` when the deferred code was run as the
scope expired.  This caused the containers within the containers set to
reorder, seemingly at random, losing the load order as specified by the
user, potentially causing script compilation issues, specifically with
user-defined files that shadow big-sourced files.